Experiement with google docs structure
Add a onscroll event handler to kix-appview-editor.So when the document scrolls
We get a callback. The callback triggers a timeout, so that we know when the user
has stopped scrolling. Use scrollTop attribute of the editor to find out how much the
user has scrolled. Dividing the scrollTop with pageHeight gives you the page user is
on. Use 'kix-paragraphrenderer' class to find out how many paragraphs are present.
Use the  'kix-lineview' to grab lines in the paragraph.