[infini-scroll] Add infini-scroll element
Add an element which can wrap another element in a scrollable
area. When the wrapped element is shorter than the infini-scroll,
show the "get more content" button; if the wrapped element
is taller than the infini-scroll, hide this button and just fire
a "get more content" event when the end of scroll is reached.