[image-detail] Add a image-detail element
Add a image-detail elements which has the following attributes:
src - Image source
title - Title for the image. default:'Image title'.
size - object {width:<>,height:<>} size of the image.
source - the web source for the image (could be different from the src).
To use this element use the following syntax :
<image-detail src=<img src> title=<img title> size={width:<>,height:<>} source=<web source> ></image-detail>

Added a functional test(image-detail-test) for this element.