[cp-scroll] Only fire one scrollend event at a time
Only fire scrollend if the user is in the "scroll gap", they
are scrolling down, and a scrollend event hasn't fired since
the scroll gap was first entered.

This prevents lots of scrollend events being fired unless the
user makes a real effort to make it happen by wiggling the
scroller up and down.