[cp-search] Abstract search result handling
Add a cp-search component which tracks the results already
retrieved and returns them in batches on demand.

The intention is that this can be used as a base for other
search implementations, which won't have to be concerned with
handling the search offset for the next search etc.

Also added deduplication code to the fetcher to ensure that
any results added to the cp-display are unique.