[cp-swapper] Integrate swapper with cp-panel and search results
NB some of these changes are because the thumbnail shown in
the gallery is now actually a thumbnail, but the dimensions shown
are those of the full-size image. This needed some changes
to manually set the width and height displayed on the tile
to the width and height of the full-size image (previously the
width and height shown were those of the image loaded into
the tile).

But the image detail panel which slides in is displaying the
full-size image, not the thumbnail.