[cp-panel] Change slider range based on number of keywords
Each time the filter is set on cp-display, change the slider's
range to reflect the number of keywords in the filter.

The keywords used by the filter are now set in cp-display, based
on the slider position.

Also add steps and a pin to the slider so the user can see
how many keywords they are displaying results for.