[cp-main] Workaround for poor arrow positioning on Mac Chrome
A bug seems to affect Chrome on Mac only, where the arrow inside
paper-button elements is misaligned vertically. This doesn't
occur on Chrome on Linux.

My guess is that it's this bug:
https://github.com/Polymer/paper-button/issues/7

The core-icon inside paper-button has "vertical-align:middle"
set, which may be causing this bug on Macs.

Overriding this CSS rule does appear to improve the vertical
alignment of the arrow on Chrome on Linux, and should
hopefully fix the issue on Macs.

(NB will only have a noticeable effect on Chrome on Mac.)

Fixes https://jira01.devtools.intel.com/browse/SE-988