[cp-tile-file] Hard-code frame for cp-tile-file
Make it possible to hard-code the frame used for particular
types of file in the layout, rather than choosing a frame
based on image size.

NB this also has the side-effect of caching the frame choice on
the tile, which should mildly improve performance when we
redo the layout (as the frame doesn't need to be chosen again).