Add test for cp-layout-generator
Added a test because the layout generator appeared to be
putting multiple items into the same part of the grid, so
they were overlapping each other.

It turned out that the problem was not the layout generator
but due to CSS: cp-tile-article did not have "position:absolute".

But I kept the test because it's always good to have more tests.