Fix site-specific searching
We need to use a siteSearch parameter on the URLs we send to
the Google search API, rather than "site:<site>".

We should also pass the sourceURL with the request, rather than
setting it as an instance variable on the search component. If
it's set on the instance, we run the risk of the incorrect site
being added to the URL (the source URL is request-specific, not
instance-specific).