[cp-search-google] Use image link as source
Using contextLink as the source for Google image searches meant
that image results which came from a page (e.g. a picture
of the Battle of Hastings from the main Wikipedia page for
Battle of Hastings) would prevent that page being used as an
article result (due to the de-duplication done by cp-fetcher).

Use the image link as the source instead.