[sources] Fix header font sizes and back button position
NB we have no visual design for the source groups screen, so
I followed the design for the main screen quite closely.

Because the source selector screen has different components
("Add" button) which clash with the main screen colours and
design, I have left that as is for now, albeit with some tweaks
to the toolbar shadow and font sizes.