Associate a tile with all keywords which apply to it
We were removing duplicate results before we had recorded
any keywords on those results against existing tiles.

This patch moves the de-duplication by source so that it occurs
after we've had a chance to harvest any keywords on results. This
means that if a result is a duplicate, but was retrieved for
a different keyword search, we can add the keyword to the existing
tile.

In turn, this means that a tile with multiple associated
keywords will now highlight all of those keywords on hover.

NB I noticed a big piece of unused code while I was doing this
(in cp-layout), so I removed it.