[cp-display] Assign source to file tiles on creation
Because the source wasn't being set on file tiles, the first tile
loaded had a source of "undefined". Then, any subsequent tile
loaded by the display would also have a source "undefined"; which
would cause it to be rejected by the deduplicator (which uses
source for the comparison between tiles).