Prevent multiple searches for the same keywords
* Only consider the first five keywords.
* Get a proper diff between keywords when keywordlist event
is handled.

This enables proper handling of search being triggered: if it
is only triggered once, we can correctly tell when the tiles
for the current set of searches have loaded. Previously, two
sets of searches were being carried out per context, which meant
that the results would load once then load again.