[cp-layout] Use an intelligent TileCache to speed up layout
Instead of appending all tiles to the layout, only append the
ones which are active. This reduces the number of tiles
which have to be considered when determining the layout,
restricting it to the set of tiles which satisfy the current
filter instead.

When the layout is calculated, any tiles which are no longer
active are removed and returned to the cache. Only the active
tiles stay as children of the container.

NB this means that when the filter changes, any tiles which are
added to the layout will appear under tiles already in the layout,
as those tiles are already children of the container.