[cp-layout] Use just two tile sizes
Reduce the choice of tile sizes to two: square and landscape.

Randomly pick one of these for image tiles; use square for
file tiles and landscape for article tiles.

This makes the layout faster, as there are fewer columns and
rows to organise tiles into. It also makes the height of all
rows the same.

However, I haven't implemented the 0.75 width / 1.25 width tiles
as shown in the design. I don't know how to get this to work with
fixed size tiles in an elegant way. If we want to have rows containing
only images with the proportions shown on the design, in combination
with article and file tiles, we're going to need a different layout
algorithm.