[cp-display] Slider now sets the 'width' on the KeywordSelector
The slider's value now ranges from 0-1 and sets a "width" on
the KeywordSelector. The width determines how many of the
active keywords (associated with the visible context) are
considered "current", i.e. should be used to filter and search
for results.

Note that the range of the slider no longer has any correspondence
with the number of keywords: it always produces values between 0
and 1.